Completes the Backbone.View based textparagraph, test-first.
Both uniteditors pass now and also the authormodel command line mocha tests.
I'm not too happy about the webpack+mocha runner in browser, particularily exception handling.
If any test throws an exception the entire test run fails and you get a line number from the bundle, with no trace of the inline-source-map.
If a test that expects an exception gets a different type of error you don't even see it in the console, you just get a failure.