Action menu can now add new units with uniteditors (to btguidline demo).
Adds additional specs on unit editors, some of which we log warnings on in editor.
Actions are now independent Command pattern objects that are typically added to context,
from any interested party such as the originating unit editor, the content model or the editor.
Issue: enter key press doesn't work in new editors.